# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ajosave API
  description: |
    Trustless rotating savings circles (Ajo/Esusu) on the Stellar blockchain.

    ## API Versioning
    This API uses URL-based versioning. All endpoints are prefixed with `/api/v1/`.
    
    **Current version:** v1  
    **Base URL:** `/api/v1/`

    Legacy endpoints without version prefix (`/api/`) are deprecated and redirect to v1.

    ## Authentication
    Most endpoints require a valid session cookie obtained via the OTP login flow:
    1. `POST /api/v1/auth/send-otp` – request a 6-digit OTP to your phone
    2. `POST /api/v1/auth/verify-otp` – exchange OTP for a session cookie (NextAuth)

    The session cookie is automatically sent by the browser on subsequent requests.
    Server-to-server callers must include the `Cookie` header.

    The cron endpoint uses a separate `Authorization: Bearer <CRON_SECRET>` header.

    ## Changelog
    - **1.1.0** – Documented `/auth/verify-otp`, `/auth/refresh`, `/profile`,
      `/sessions`, `/fx/rate`, `/kyc/verify`, and `/circles/{id}/leave`.
      This is not the full route surface (see `src/app/api/v1/` for every
      handler); coverage will keep expanding incrementally per endpoint.
    - **1.0.0** – Initial versioned spec covering auth, circles, cron, health.
  version: "1.1.0"
  contact:
    email: security@ajosave.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://www.ajosave.app/api/v1
    description: Production
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie set after OTP verification
    cronBearer:
      type: http
      scheme: bearer
      description: Static secret from `CRON_SECRET` env var
  schemas:
...
    ApiSuccess:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          description: Payload – see individual endpoint schemas
    ApiError:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
          example: "Circle not found"
        code:
          type: string
          example: "UNAUTHORIZED"

    # ── Domain objects ──────────────────────────────────────────────────────────
...
  responses:
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            success: false
            error: "Too many requests. Please wait before requesting another OTP."
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            success: false
            error: "Internal server error"
            code: "INTERNAL_ERROR"

# ─── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /auth/send-otp:
    post:
      tags: [auth]
      summary: Request a phone OTP
      description: |
        Sends a 6-digit OTP via SMS (Termii). Rate-limited to **3 requests per 10 minutes** per phone number.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  description: E.164 or local format (10–15 digits)
                  example: "+2348012345678"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "OTP sent successfully"
        "400":
          description: Invalid phone number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate a refresh token
      description: Exchanges a valid refresh token for a new access token and refresh token pair.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: New token pair issued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accessToken:
                            type: string
                          refreshToken:
                            type: string
                          expiresIn:
                            type: integer
                            example: 900
        "400":
          description: Missing refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...